AO3-5392 Rewrite set_revised_at_by_chapter.
Comparisons between dates and times aren't guaranteed to work properly,
so we want to rewrite set_revised_at_by_chapter to avoid all such
comparisons. This should help address cases where it's impossible to
change the date of a backdated work.

We also want to avoid setting revised_at to a date in the future (which
can happen because set_revised_at converts dates to noon UTC time).